some powerup adjustments started mech logic

+++ Assets/Resources/Scripts/IOnDamage.cs
@@ -2,6 +2,6 @@ namespace LaninCode
 {
     public interface IOnDamage
     {
-        void SetHealth(float health);
+        void SetHealth(int health);
     }
 }
